+++ src/modules/auth/components/signIn/EmailInput.tsx
@@ -4,7 +4,7 @@ import { useTranslation } from 'react-i18next';
 import { NS } from '@/config/constants';
 
 import { StyledTextField } from '../common/StyledTextField';
-import { EmailAdornment } from '../common/adornments';
+import { EmailAdornment } from '../common/adornment';
 
 type Props = {
   id?: string;
